Implemented attack concurrently

+++ Age-of-Concurency/src/com/upt/IAttacker.java
@@ -3,5 +3,5 @@ package com.upt;
 public interface IAttacker
 {
     void attack(IAttacker attacker);
-    void takeDamage(float damage);
+    void takeDamage(int damage);
 }
